requirements.txt
pyam_iamc
streamlit
pandas

